Refactor state module
We split out the actual state resolution algorithm to prepare for having
multiple versions.
